+++ src/draggers/SoTransformerDragger.cpp
@@ -709,7 +709,7 @@ SoTransformerDragger::SoTransformerDragger(void)
   if (SO_KIT_IS_FIRST_INSTANCE()) {
     SoInteractionKit::readDefaultParts("transformerDragger.iv",
                                        TRANSFORMERDRAGGER_draggergeometry,
-                                       strlen(TRANSFORMERDRAGGER_draggergeometry));
+                                       (int)strlen(TRANSFORMERDRAGGER_draggergeometry));
   }
 
   SO_KIT_ADD_FIELD(rotation, (SbRotation(SbVec3f(0.0f, 0.0f, 1.0f), 0.0f)));
