+++ src/draggers/SoTranslate1Dragger.cpp
@@ -157,7 +157,7 @@ SoTranslate1Dragger::SoTranslate1Dragger(void)
   if (SO_KIT_IS_FIRST_INSTANCE()) {
     SoInteractionKit::readDefaultParts("translate1Dragger.iv",
                                        TRANSLATE1DRAGGER_draggergeometry,
-                                       strlen(TRANSLATE1DRAGGER_draggergeometry));
+                                       (int)strlen(TRANSLATE1DRAGGER_draggergeometry));
   }
 
   SO_KIT_ADD_FIELD(translation, (0.0f, 0.0f, 0.0f));
